feat(core): test the flow from parser to html deck generation
